+++ src/1-one/sweep/methods/tagger/_logger.js
@@ -1,8 +1,16 @@
+
+const isArray = function (arr) {
+  return Object.prototype.toString.call(arr) === '[object Array]'
+}
+
 const logger = function (todo, document) {
   let [n, start, end] = todo.pointer
   let terms = document[n]
   let i = start > 4 ? start - 2 : 0
-  let tag = typeof todo.tag !== 'string' ? todo.tag.join(' #') : todo.tag
+  let tag = todo.tag || ''
+  if (isArray(todo.tag)) {
+    tag = todo.tag.join(' #')
+  }
   let reason = todo.reason || todo.match
   reason = reason ? `|${reason}|` : ''
   let msg = `  ${reason}`.padEnd(20) + ' - '